requirements.txt
#python==3.10
pandas==1.3.5
numpy==1.26.4
google-genai==1.16.1
gspread==6.2.1
openpyxl==3.1.5

requests>=2.31.0
typing-extensions>=4.9.0
lark==1.2.2 
scikit-learn
PyPDF2


gspread
google-auth-oauthlib
oauth2client

tk
